==> src/google_trend/00_google_trend_data_preparation_FOR_REFERENCE.R <==


readRawdata2.1 <- function(BASEDIR){
  # purpose: 観光地名検索量データ取得
  # args: 
  #   BASEDIR: ベース
  # return: 
  #   a list
  # notes:
  #   WindowsではgtrendsRパッケージがうまく動かない
  
  library(gtrendsR)
  library(stringi)
  
  # - - - - -
  sFILENAME = "citylist_2.csv"
  # - - - - -
  
  cat("[readRawdata2.1] Hello\n")
  
  # category 67:Travel
  # print(categories %>% filter(grepl("Travel", name))
  
  # country_code JP: JAPAN
  # print(countries %>% filter(grepl("JAPAN", name)))
  
  dfCity <- read.csv(
    paste0(BASEDIR, "/", sFILENAME),
    stringsAsFactors = FALSE
  ) 
  
  lOut <- lapply(
    dfCity$sCity,
    function(sCurrent){
      cat("[readRawdata2.1]", sCurrent, "...\n")
      Sys.setlocale(locale="C")
      out <- try(
        gtrends(
          keyword  = sCurrent,
          geo      = "JP",
          time     = "2014-01-01 2018-12-31",
          category = 67
        )
      )
      Sys.setlocale("LC_ALL", 'ja')
      return(out)
    }
  )
  cat("[readRawdata2.1] Done.\n")
  return(lOut)
}


makeData2.1 <- function(lRawdata2){
  # purpose: 観光地名検索量データ2の整形
  # args: 
  #   dfRawdata: generated by readRawdata2.1()
  # return: 
  #   a data frame
  #   行: 週(2014-2018)
  #   $nCityID
  #   $keyword
  #   $date
  #   $hit
  
  cat("[makeData2.1] Hi,\n")
  
  lOut <- lapply(
    seq_along(lRawdata2), 
    function(nCityID){
      out <- lRawdata2[[nCityID]]$interest_over_time
      if (!is.null(out)){
        out <- out %>%
          mutate(
            nCityID = nCityID,
            keyword = iconv(keyword,from="utf8",to="shift_jis")
          )
      }
      return(out)
    }
  )
  abExist <- sapply(lOut, function(x) {!is.null(x)})
  out <- bind_rows(lOut[abExist]) %>%
    dplyr::select(nCityID, date, hits, keyword) %>%
    arrange(nCityID, date) %>%
    group_by(nCityID) %>%
    mutate(
      gSeasonal = as.vector(
        decompose(ts(hits, start=c(2014, 1), frequency = 52), type = "additive")$seasonal  
      ), 
      gAdjusted = hits - gSeasonal
    )
  
  cat("[makeData2.1] Done\n")
  return(out)
}


exportData2.1 <- function(dfData2, WORKDIR){
  # purpose: 観光地名検索量データのCSV出力 (書籍原稿用)
  # args: 
  #   dfData2: generated by makeData2.1()
  #   WORKDIR: 出力先
  
  write.csv(dfData2, paste0(WORKDIR, "/dfData2.csv"), row.names = FALSE)
  
}


### 参考: データの取得と整形
### s.lRawdata2 <- readRawdata2.1 (WORKDIR)
### s.dfData2   <- makeData2.1 (s.lRawdata2)
### exportData2.1(s.dfData2, WORKDIR)
