==> src/google_trend/03_google_trend_dynamic_factor_analysis_by_rstan_02_rotation.R <==
# setwd("//media//kswada//MyFiles//R//google_trend//")
setwd("//media//kswada//MyFiles//R//Bayesian_inference//google_trend//")

packages <- c("dplyr", "rstan")
purrr::walk(packages, library, character.only = TRUE, warn.conflicts = FALSE)



# ------------------------------------------------------------------------------
# data:  google trend
# ------------------------------------------------------------------------------

library(tidyverse)

s.dfData2 <- read_csv("dfData2.csv", locale = locale(encoding = "cp932"))


str(s.dfData2)


dim(s.dfData2)


car::some(s.dfData2)



# ------------------------------------------------------------------------------
# Dynamic Fdactor Analysis:  rotation function
# ------------------------------------------------------------------------------

sub_Rotate_Stan.1 <- function(lModel){
  # purpose: 
  #   DFAの因子を回転する
  # args:
  #   lModel: generated by Sub_Estimate_Stan.*()
  # return: 
  #   a list of matrices
  
  cat("[sub_Rotate_Stan.1] Hello\n")
  
  # nNumMCMC, nNumSeries, nNumFactor
  mgZ <- rstan::extract(lModel)$mgZ
  
  # nNumMCMC, nNumFactor, nNumTime
  mgX <- rstan::extract(lModel)$mgX
  
  cat("[sub_Rotate_Stan.1] mean -> rotation ...\n")
  
  mgZ_mean <- apply(mgZ, c(2, 3), mean)
  mgX_mean <- apply(mgX, c(2, 3), mean)
  mgRotMat <- stats::varimax(mgZ_mean)$rotmat
  mgZ_mean_rotated <- mgZ_mean %*% mgRotMat
  mgX_mean_rotated <- solve(mgRotMat) %*% mgX_mean
  
  cat("[sub_Rotate_Stan.1] rotation -> mean ...\n")
  
  nNumMCMC   <- dim(mgZ)[1]
  nNumSeries <- dim(mgZ)[2]
  nNumFactor <- dim(mgZ)[3]
  nNumTime   <- dim(mgX)[3]
  
  mgZ_rotated <- array(0, dim = c(nNumMCMC, nNumSeries, nNumFactor))
  mgX_rotated <- array(0, dim = c(nNumMCMC, nNumFactor, nNumTime))
  
  for(i in 1:nNumMCMC) {
    mgCurrentZ <- mgZ[i,,]
    mgCurrentX <- mgX[i,,]
    
    mgRotMat <- stats::varimax(mgCurrentZ)$rotmat
    mgCurrentZ_rotated <- mgCurrentZ %*% mgRotMat
    mgCurrentX_rotated <- solve(mgRotMat) %*% mgCurrentX
    
    mgZ_rotated[i,,] <- mgCurrentZ_rotated
    mgX_rotated[i,,] <- mgCurrentX_rotated
  }
  
  mgZ_rotated_mean <- apply(mgZ_rotated, c(2,3), mean)
  mgX_rotated_mean <- apply(mgX_rotated, c(2,3), mean)
  
  lOut <- list(
    mgZ_mean         = mgZ_mean, 
    mgX_mean         = mgX_mean, 
    mgZ_mean_rotated = mgZ_mean_rotated, 
    mgX_mean_rotated = mgX_mean_rotated, 
    mgZ_rotated      = mgZ_rotated, 
    mgX_rotated      = mgX_rotated, 
    mgZ_rotated_mean = mgZ_rotated_mean,
    mgX_rotated_mean = mgX_rotated_mean
  )
  cat("[sub_Rotate_Stan.1] Done.\n")
  return(lOut)
}



# ------------------------------------------------------------------------------
# Dynamic Fdactor Analysis:  rotation
# ------------------------------------------------------------------------------

s.lDFARot108 <- sub_Rotate_Stan.1(s.lDFA108)
